Moving handleSeedPhraseSubmit to RecoveryFooter to avoid props drilling
